module_3(fix): const lists refactoring
