+++ src/components/footer/footer-navigation/footer-navigation-items.const.ts
@@ -1,0 +1,52 @@
+export const FOOTER_NAVIGATION_ITEMS = [
+  {
+    title: "About bank",
+    to: "#",
+    id: "001",
+  },
+  {
+    title: "Ask a Question",
+    to: "#",
+    id: "002",
+  },
+  {
+    title: "Quality of service",
+    to: "#",
+    id: "003",
+  },
+  {
+    title: "Requisites",
+    to: "#",
+    id: "004",
+  },
+  {
+    title: "Press center",
+    to: "#",
+    id: "005",
+  },
+  {
+    title: "Bank career",
+    to: "#",
+    id: "006",
+  },
+  {
+    title: "Investors",
+    to: "#",
+    id: "007",
+  },
+  {
+    title: "Analytics",
+    to: "#",
+    id: "008",
+  },
+  {
+    title: "Business and processes",
+    to: "#",
+    id: "009",
+  },
+  {
+    title: "Compliance and business ethics",
+    to: "#",
+    id: "010",
+  },
+];
