+++ src/components/footer/footer-navigation/footer-navigation.tsx
@@ -1,54 +1,12 @@
+import { FOOTER_NAVIGATION_ITEMS } from "./footer-navigation-items.const";
 import styles from "./footer-navigation.module.scss";
-
-const links = [
-  {
-    title: "About bank",
-    to: "#",
-  },
-  {
-    title: "Ask a Question",
-    to: "#",
-  },
-  {
-    title: "Quality of service",
-    to: "#",
-  },
-  {
-    title: "Requisites",
-    to: "#",
-  },
-  {
-    title: "Press center",
-    to: "#",
-  },
-  {
-    title: "Bank career",
-    to: "#",
-  },
-  {
-    title: "Investors",
-    to: "#",
-  },
-  {
-    title: "Analytics",
-    to: "#",
-  },
-  {
-    title: "Business and processes",
-    to: "#",
-  },
-  {
-    title: "Compliance and business ethics",
-    to: "#",
-  },
-];
 
 export const FooterNavigation = () => {
   return (
     <nav>
       <ul className={styles.navigation}>
-        {links.map(item => (
-          <li key={item.title} className={styles.navigation__item}>
+        {FOOTER_NAVIGATION_ITEMS.map(item => (
+          <li key={item.id} className={styles.navigation__item}>
             <a href={item.to} target="_blank">
               {item.title}
             </a>
