+++ src/components/header/header-navigation/header-navigation-items.const.ts
@@ -1,0 +1,22 @@
+export const HEADER_NAVIGATION_ITEMS = [
+  {
+    title: "Credit card",
+    to: "/credit-card",
+    id: "001",
+  },
+  {
+    title: "Product",
+    to: "/product",
+    id: "002",
+  },
+  {
+    title: "Account",
+    to: "/account",
+    id: "003",
+  },
+  {
+    title: "Resources",
+    to: "/resources",
+    id: "004",
+  },
+];
