+++ src/components/page-blocks/credit-card-page-blocks/credit-card-tabs/components/cashback/cashback-items.const.ts
@@ -2,25 +2,31 @@ export const CASHBACK_ITEMS = [
   {
     title: "5%",
     content: "For food delivery, cafes and restaurants",
+    id: "001",
   },
   {
     title: "5%",
     content: "In supermarkets with our subscription",
+    id: "002",
   },
   {
     title: "2%",
     content: "In clothing stores and children's goods",
+    id: "003",
   },
   {
     title: "1%",
     content: "Other purchases and payment of services and fines",
+    id: "004",
   },
   {
     title: "up to 3%",
     content: "Shopping in online stores",
+    id: "005",
   },
   {
     title: "30%",
     content: "Purchases from our partners",
+    id: "006",
   },
 ];
