+++ src/components/page-blocks/credit-card-page-blocks/credit-card-tabs/components/rates-conditions/rates-conditions-items.const.ts
@@ -2,27 +2,33 @@ export const RATES_CONDITIONS_ITEMS = [
   {
     title: "Card currency",
     content: "Rubles, dollars, euro",
+    id: "001",
   },
   {
     title: "Interest free period",
     content: "0% up to 160 days",
+    id: "002",
   },
   {
     title: "Payment system",
     content: "Mastercard, Visa",
+    id: "003",
   },
   {
     title: "Maximum credit limit on the card",
     content: "600 000 ₽",
+    id: "004",
   },
   {
     title: "Replenishment and withdrawal",
     content:
       "At any ATM. Top up your credit card for free with cash or transfer from other cards",
+    id: "005",
   },
   {
     title: "Max cashback per month",
     content: "15 000 ₽",
+    id: "006",
   },
   {
     title: "Transaction Alert",
@@ -30,5 +36,6 @@ export const RATES_CONDITIONS_ITEMS = [
       "60 ₽ — SMS or push notifications",
       "0 ₽ — card statement, information about transactions in the online bank",
     ],
+    id: "007",
   },
 ];
