+++ src/components/page-blocks/credit-card-page-blocks/credit-card-tabs/credit-card-tabs-items.const.tsx
@@ -6,7 +6,7 @@ export type CreditCardTabT = {
   link: string;
 };
 
-export const TABS_LIST = [
+export const CREDIT_CARD_TABS_ITEMS = [
   {
     title: "About card",
     content: <AboutCard />,
