+++ src/components/page-blocks/credit-card-page-blocks/form/components/form-title/form-title.module.scss
@@ -15,6 +15,6 @@
 }
 
 .form__heading {
-  @include mixins.secondaryHeading;
   font-weight: 700;
+  @include mixins.secondaryHeading;
 }
