+++ src/components/page-blocks/credit-card-page-blocks/get-card/get-card.tsx
@@ -1,4 +1,5 @@
-import { InfoBlock, INFO_BLOCK_ITEMS } from "./info-block";
+import { InfoBlock } from "./info-block";
+import { GET_CARD_ITEMS } from "./get-card-items.const";
 import styles from "./get-card.module.scss";
 
 export const GetCard = () => {
@@ -6,7 +7,7 @@ export const GetCard = () => {
     <section className={styles.wrapper}>
       <h2 className={styles.heading}>How to get a card</h2>
       <ul className={styles.list}>
-        {INFO_BLOCK_ITEMS.map((item, index) => (
+        {GET_CARD_ITEMS.map((item, index) => (
           <li key={item.id} className={styles.list__item}>
             <InfoBlock num={++index} text={item.text} />
           </li>
