+++ src/components/page-blocks/credit-card-page-blocks/get-card/index.ts
@@ -1,0 +1,1 @@
+export * from "./get-card";
