+++ src/components/page-blocks/credit-card-page-blocks/get-card/info-block/index.ts
@@ -1,2 +1,1 @@
 export * from "./info-block";
-export * from "./info-block-items.const";
